Refactor frontend to use modern module-based architecture
This commit addresses several legacy practices in the frontend code to align with modern web development standards.

Key changes include:
- Migrated all inline JavaScript from `index.html` into a new, single entry point at `src/main.js`.
- Replaced all CDN-based library links (GSAP, Chart.js, Flatpickr, FontAwesome) with dependencies managed by `pnpm`. The application now imports these libraries as ES modules.
- Integrated Tailwind CSS directly into the Vite build process, removing the CDN script and enabling optimizations like style purging.
- Removed numerous redundant and obsolete files, including old versions of JS application logic (`app.js`, `script.js`), PHP API endpoints, and backup/optimized JSON files.
- Added a baseline ESLint configuration (`.eslintrc.cjs`) to enforce code quality.
- Cleaned up `index.html` to be a clean application shell, loading only the single JS module.

These changes significantly improve the project's structure, maintainability, and performance by leveraging the capabilities of the Vite build tool.
