+++ index.html
@@ -9,31 +9,10 @@
     <meta property="og:title" content="Radio Adamowo – Przewodnik po Toksycznych Relacjach">
     <meta property="og:description" content="Analizujemy rodzinną historię jako studium przypadku manipulacji, narcyzmu i toksycznych zachowań.">
     <meta property="og:type" content="website">
-
     <link rel="icon" href="/images/icons/favicon.svg">
-    
-    <script src="https://cdn.tailwindcss.com"></script>
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700&family=Special+Elite&family=Creepster&display=swap" rel="stylesheet">
-    
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/flatpickr/dist/flatpickr.min.css">
-    <link rel="stylesheet" href="https://npmcdn.com/flatpickr/dist/themes/dark.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
-
-    <!-- GSAP for advanced animations -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.5/gsap.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.5/Draggable.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.5/MotionPathPlugin.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.5/ScrollTrigger.min.js"></script>
-
-    <!-- Chart.js for data visualization -->
-    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-
-    <!-- HLS.js for live streaming support -->
-    <script src="https://cdn.jsdelivr.net/npm/hls.js@latest/dist/hls.min.js"></script>
-
-    <link rel="stylesheet" href="./style.css">
 </head>
 <body class="antialiased bg-wave-pattern">
     <!-- Loading Screen -->
@@ -950,107 +929,6 @@
         </div>
     </div>
     
-    <!-- Scripts -->
-    <script src="https://cdn.jsdelivr.net/npm/flatpickr"></script>
-    <script src="https://npmcdn.com/flatpickr/dist/l10n/pl.js"></script>
-    
-    <!-- Plugin System (New Architecture) -->
-    <script src="./src/plugins.js"></script>
-    
-    <!-- Backward Compatibility: Original Script -->
-    <script src="./script.js"></script>
-    
-    <!-- Enhanced Integration Script -->
-    <script>
-        // Enhanced initialization with plugin system integration
-        document.addEventListener('DOMContentLoaded', async () => {
-            // Wait for plugin manager to initialize
-            if (window.pluginManager) {
-                // Hook into the original script functionality
-                window.pluginManager.addHook('audio.play', (audioElement) => {
-                    console.log('Plugin system: Audio play event triggered');
-                    return audioElement;
-                });
-
-                window.pluginManager.addHook('comment.added', (commentData) => {
-                    console.log('Plugin system: Comment added event triggered', commentData);
-                    return commentData;
-                });
-
-                console.log('✓ Radio Adamowo Extended - Plugin system integrated');
-                console.log('✓ Backward compatibility maintained');
-                console.log('✓ All original features preserved');
-            }
-
-            // Enhanced API integration
-            if (typeof RadioAdamowo !== 'undefined') {
-                // Extend original functionality without breaking it
-                const originalFunctions = RadioAdamowo || {};
-                
-                // Add API endpoints for mobile/desktop synchronization
-                window.RadioAdamowoAPI = {
-                    // Stream API
-                    stream: {
-                        getStatus: () => fetch('/api/v1/stream/status').then(r => r.json()),
-                        getCurrent: () => fetch('/api/v1/stream/current').then(r => r.json()),
-                        getPlaylist: (page = 1, limit = 20) => 
-                            fetch(`/api/v1/stream/playlist?page=${page}&limit=${limit}`).then(r => r.json())
-                    },
-                    
-                    // Comments API (Enhanced)
-                    comments: {
-                        list: (page = 1, limit = 20) => 
-                            fetch(`/api/v1/comments?page=${page}&limit=${limit}`).then(r => r.json()),
-                        add: (comment) => 
-                            fetch('/api/v1/comments/add', {
-                                method: 'POST',
-                                headers: { 'Content-Type': 'application/json' },
-                                body: JSON.stringify(comment)
-                            }).then(r => r.json()),
-                        getByDate: (date) => 
-                            fetch(`/api/v1/comments/date?date=${date}`).then(r => r.json()),
-                        react: (commentId, type) => 
-                            fetch(`/api/v1/comments/react/${commentId}`, {
-                                method: 'POST',
-                                headers: { 'Content-Type': 'application/json' },
-                                body: JSON.stringify({ type })
-                            }).then(r => r.json())
-                    },
-                    
-                    // Notifications API
-                    notifications: {
-                        getUnread: () => fetch('/api/v1/notifications/unread').then(r => r.json()),
-                        markAsRead: (id) => 
-                            fetch(`/api/v1/notifications/read/${id}`, { method: 'PUT' }).then(r => r.json()),
-                        subscribe: (subscription) => 
-                            fetch('/api/v1/notifications/subscribe', {
-                                method: 'POST',
-                                headers: { 'Content-Type': 'application/json' },
-                                body: JSON.stringify(subscription)
-                            }).then(r => r.json())
-                    }
-                };
-
-                console.log('✓ Enhanced API endpoints available at window.RadioAdamowoAPI');
-            }
-        });
-
-        // Service Worker registration for enhanced PWA features
-        if ('serviceWorker' in navigator) {
-            navigator.serviceWorker.register('/sw-comprehensive.js')
-                .then(registration => {
-                    console.log('✓ Enhanced Service Worker registered:', registration);
-                    
-                    // Listen for push notifications
-                    registration.addEventListener('pushsubscriptionchange', event => {
-                        console.log('Push subscription changed');
-                        // Re-subscribe logic here
-                    });
-                })
-                .catch(error => {
-                    console.warn('Service Worker registration failed:', error);
-                });
-        }
-    </script>
+    <script type="module" src="/src/main.js"></script>
 </body>
 </html>
