+++ package.json
@@ -17,14 +17,14 @@
     "lighthouse": "lhci autorun"
   },
   "keywords": [
-    "radio", 
-    "pwa", 
-    "streaming", 
-    "education", 
-    "psychology", 
-    "manipulation", 
-    "security", 
-    "web-audio", 
+    "radio",
+    "pwa",
+    "streaming",
+    "education",
+    "psychology",
+    "manipulation",
+    "security",
+    "web-audio",
     "hls"
   ],
   "author": "Radio Adamowo Team <contact@radioadamowo.pl>",
@@ -43,21 +43,25 @@
     "npm": ">=8.0.0"
   },
   "dependencies": {
-    "hls.js": "^1.5.13",
-    "chart.js": "^4.4.1",
-    "gsap": "^3.12.5"
+    "@fortawesome/fontawesome-free": "^7.0.1",
+    "chart.js": "^4.5.0",
+    "flatpickr": "^4.6.13",
+    "gsap": "^3.13.0",
+    "hls.js": "^1.6.13"
   },
   "devDependencies": {
-    "vite": "^5.4.8",
-    "vite-plugin-pwa": "^0.20.1",
+    "@tailwindcss/vite": "^4.1.13",
+    "@vitest/coverage-v8": "^1.6.0",
+    "autoprefixer": "^10.4.20",
+    "cssnano": "^6.1.2",
     "eslint": "^8.57.1",
     "prettier": "^3.3.3",
+    "tailwindcss": "^4.1.13",
     "typescript": "^5.6.3",
+    "vite": "^5.4.8",
+    "vite-plugin-pwa": "^0.20.1",
     "vitest": "^1.6.0",
-    "@vitest/coverage-v8": "^1.6.0",
-    "workbox-build": "^7.0.0",
-    "autoprefixer": "^10.4.20",
-    "cssnano": "^6.1.2"
+    "workbox-build": "^7.0.0"
   },
   "browserslist": [
     "> 1%",
@@ -77,4 +81,3 @@
     "start_url": "/"
   }
 }
-
