+++ tailwind.config.js
@@ -1,0 +1,12 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
+    "./level2/**/*.html",
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
