+++ vite.config.js
@@ -1,6 +1,10 @@
 import { defineConfig } from 'vite';
+import tailwindcss from '@tailwindcss/vite';
 
 export default defineConfig({
+  plugins: [
+    tailwindcss(),
+  ],
   server: {
     port: 3000,
     host: true,
